+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - No more weird notes in the release description
 - Basic logic for new libary handling system
 - Support for 7-Zip
+- `--confirm-dl` argument. It will automaticly download and install new drivers! (uses minimal installer)
 
 ### Changed
 - Updated HtmlAgilityPack to the latest version